refactor: MenuProducts 일급컬렉션 분리
